[feat]: DTO 수정, repository, viewmodel 기본 틀 작성
