--- app/src/main/java/com/example/durymong/model/dto/response/column/CategoryResponse.kt
@@ -1,16 +1,0 @@
-package com.example.durymong.model.dto.response.column
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CategoryResponse(
-    val code: Int,
-    val message: String,
-    val result : List<Category>
-)
-
-@Serializable
-data class Category(
-    val name: String,
-    val image: String
-)
