+++ app/src/main/java/com/example/durymong/model/dto/response/column/CategoryResponseDto.kt
@@ -1,0 +1,18 @@
+package com.example.durymong.model.dto.response.column
+
+
+data class CategoryResponseDto(
+    val success: Boolean,
+    val code: Int,
+    val message: String,
+    val result : CategoryResult
+)
+
+data class CategoryResult(
+    val categories: List<Category> // categories 배열을 담는 객체
+)
+
+data class Category(
+    val name: String,
+    val image: String
+)
