+++ app/src/main/java/com/example/durymong/retrofit/service/ColumnService.kt
@@ -1,4 +1,10 @@
 package com.example.durymong.retrofit.service
 
+import com.example.durymong.model.dto.response.column.CategoryResponseDto
+import retrofit2.Call
+import retrofit2.http.GET
+
 interface ColumnService {
+    @GET("column")
+    fun getColumnCategories(): Call<CategoryResponseDto>
 }
